requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
Pillow>=10.0,<11.0
python-decouple>=3.8
whitenoise>=6.6.0
django-cors-headers>=4.3

#pip install -r requirements.txt
